% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{BuildPCSFNet}
\alias{BuildPCSFNet}
\title{Compute minimum connect subnetwork based on input nodes using prize-collecting steiner forest approach}
\usage{
BuildPCSFNet(dataSetObj = NA)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}
}
\value{
check overall.graph object for result
}
\description{
Compute minimum connect subnetwork based on input nodes using prize-collecting steiner forest approach
}
