% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{BuildSeedProteinNet}
\alias{BuildSeedProteinNet}
\title{Create network from only input (seeds)}
\usage{
BuildSeedProteinNet(dataSetObj = NA)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}
}
\description{
Create network from only input (seeds)
}
