% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{call_sr}
\alias{call_sr}
\title{Internal function \code{call_sr}}
\usage{
call_sr(from, to, cost, node_names, node_prizes)
}
\arguments{
\item{from}{A \code{CharacterVector} that corresponds to \code{head} nodes of the edges.}

\item{to}{A \code{CharacterVector} that corresponds the \code{tail} nodes of the edges.}

\item{cost}{A \code{NumericVector} which represents the edge weights.}

\item{node_names}{A \code{CharacterVector} demonstrates the names of the nodes.}

\item{node_prizes}{A \code{NumericVector} which corresponds to the node prizes.}
}
\description{
This function is internally used to solve the PCST.
}
\author{
Murodzhon Akhmedov
}
\keyword{internal}
