% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{CreateGraph}
\alias{CreateGraph}
\title{Create igraph object from edgelist created from database selection and decompose into connected subnetworks}
\usage{
CreateGraph(dataSetObj = NA)
}
\description{
Create igraph object from edgelist created from database selection and decompose into connected subnetworks
}
