% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{DoGba}
\alias{DoGba}
\title{Guilt-by-association analysis using Random Walk With Restart}
\usage{
DoGba(fileNm = "NA", method = "rwr", nodeids)
}
\arguments{
\item{fileNm}{Input file name for exporting result table as .csv}

\item{method}{Method name, only "rwr" is supported}

\item{nodeids}{The IDs of nodes set as seed nodes "; " separated.}
}
\description{
Guilt-by-association analysis using Random Walk With Restart
}
