% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{ExtractModule}
\alias{ExtractModule}
\title{Compute minimum connected subnetwork from input nodes using shortest path based approach}
\usage{
ExtractModule(dataSetObj = NA, nodeids, dim = "3")
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}

\item{nodeids}{The ids of input nodes}

\item{dim}{dimension of network results to be exported}
}
\description{
Compute minimum connected subnetwork from input nodes using shortest path based approach
}
