% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{FindCommunities}
\alias{FindCommunities}
\title{Detect graph communities using function from igraph}
\usage{
FindCommunities(method = "infomap", use.weight = FALSE)
}
\arguments{
\item{method}{algorithm choice: walktrap, infomap, label propagation}

\item{use.weight}{Boolean consider edge weight: FALSE}
}
\description{
Detect graph communities using function from igraph
}
