% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{GetRCommandHistory}
\alias{GetRCommandHistory}
\title{Export R Command History}
\usage{
GetRCommandHistory(dataSetObj = NA)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}
}
\description{
Export R Command History
}
