% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{GetShortestPaths}
\alias{GetShortestPaths}
\title{Compute the shortest path between two nodes}
\usage{
GetShortestPaths(from, to, intermediate = "false")
}
\arguments{
\item{from}{ID of from node}

\item{to}{ID of target node}

\item{intermediate}{if true check whether from and target node connects to each other}
}
\description{
Compute the shortest path between two nodes
}
