% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPeakNet.R
\name{ImportMSPeaks}
\alias{ImportMSPeaks}
\title{Importing peak file}
\usage{
ImportMSPeaks(PeakFile = NA)
}
\arguments{
\item{PeakFile}{Name of input peak file}
}
\value{

}
\description{
Importing peak file
}
