% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPeakNet.R
\name{InitiatePeakSet}
\alias{InitiatePeakSet}
\title{Title}
\usage{
InitiatePeakSet(ionMode = -1, ppm = 10, database = "KEGG", p_cutoff = 0.1)
}
\arguments{
\item{ionMode}{Positive or negative 1 is pos, -1 is neg}

\item{ppm}{PPM value}

\item{database}{Name of database (kegg, pubchem, hmdb)}

\item{p_cutoff}{P-value cutoff}
}
\value{

}
\description{
Title
}
