% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionUtils.R
\name{PerformMetEnrichment}
\alias{PerformMetEnrichment}
\title{Perform metabolite or gene/metabolite enrichment analysis}
\usage{
PerformMetEnrichment(dataSetObj = NA, file.nm, fun.type, ids)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}

\item{file.nm}{File name of result table to be exported in csv format, do not include file extension}

\item{fun.type}{Enrichment type: keggm (kegg metabolite), integ (joint gene/metabolite)}

\item{ids}{String of ids to be tested separated by "; "}
}
\description{
Perform metabolite or gene/metabolite enrichment analysis
}
