% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionUtils.R
\name{PerformNetEnrichment}
\alias{PerformNetEnrichment}
\title{Perform gene enrichment analysis or identify gene regulatory targets}
\usage{
PerformNetEnrichment(file.nm, fun.type, IDs)
}
\arguments{
\item{file.nm}{File name of result table to be exported in csv format, do not include file extension}

\item{fun.type}{Enrichment database type}

\item{IDs}{String of ids to be tested separated by "; "}
}
\description{
Perform gene enrichment analysis or identify gene regulatory targets
}
