% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{PlotDegreeHistogram}
\alias{PlotDegreeHistogram}
\title{PlotDegreeHistogram}
\usage{
PlotDegreeHistogram(imgNm, netNm = "NA", dpi = 72, format = "png")
}
\arguments{
\item{imgNm}{image name}

\item{netNm}{network name}

\item{dpi}{dpi value}

\item{format}{format}
}
\description{
PlotDegreeHistogram
}
