% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{PrepareGraph}
\alias{PrepareGraph}
\title{PrepareGraph}
\usage{
PrepareGraph(net.nm, type = "", export = T)
}
\arguments{
\item{net.nm}{net.nm}

\item{type}{type}

\item{export}{export}
}
\description{
PrepareGraph
}
