% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{PrepareNetwork}
\alias{PrepareNetwork}
\title{Prepare the json file for network visualization}
\usage{
PrepareNetwork(dataSetObj = NA, net.nm, json.nm)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}

\item{net.nm}{Name of subnetwork created from network building (i.e subnetwork1)}

\item{json.nm}{Name of json file to be exported for network visualization}
}
\description{
Prepare the json file for network visualization
}
