% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{QueryNetMulti}
\alias{QueryNetMulti}
\title{Query database using input list to compute omics network when more than one input list are uploaded}
\usage{
QueryNetMulti(
  dataSetObj = NA,
  type = "gene",
  dbType = "default",
  inputType = "gene"
)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}

\item{type}{Network type (gene, met, mir, tf, mic, peak, m2m, snp)}

\item{dbType}{Database name (i.e innatedb)}

\item{inputType}{Omics type of input features (gene, protein, met, tf, mic, peak, snp)}
}
\description{
Query database using input list to compute omics network when more than one input list are uploaded
}
