% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFileUtils.R
\name{ReadGraphFile}
\alias{ReadGraphFile}
\title{ReadGraphFile}
\usage{
ReadGraphFile(dataSetObj = NA, fileName, fileType)
}
\arguments{
\item{dataSetObj}{dataSetObj}

\item{fileName}{fileName, Input name of graph file}

\item{fileType}{fileType, File type of graph file (".json", ".graphml", ".txt", ".sif")}
}
\description{
ReadGraphFile
}
