% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{RecordRCommand}
\alias{RecordRCommand}
\title{Record R Commands}
\usage{
RecordRCommand(dataSetObj = NA, cmd)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}

\item{cmd}{Commands}
}
\description{
Record R Commands
}
