% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{resetNetwork}
\alias{resetNetwork}
\title{Resetting individual omics networks generated from database selection step}
\usage{
resetNetwork(dataSetObj = NA)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}
}
\description{
Resetting individual omics networks generated from database selection step
}
