% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{SanityCheckSelection}
\alias{SanityCheckSelection}
\title{SanityCheckSelection}
\usage{
SanityCheckSelection(dataSetObj = NA)
}
\arguments{
\item{dataSetObj}{dataSetObj}
}
\description{
SanityCheckSelection
}
