% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{SaveNetworkJson}
\alias{SaveNetworkJson}
\title{Save graph file in OmicsNet json format}
\usage{
SaveNetworkJson(fileNm)
}
\arguments{
\item{fileNm}{The file name of output json file}
}
\description{
Save graph file in OmicsNet json format
}
