% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{SetMetPotentialOpts}
\alias{SetMetPotentialOpts}
\title{SetMetPotentialOpts}
\usage{
SetMetPotentialOpts(metPotentialThresh, currExclude, uniExclude, orphExclude)
}
\arguments{
\item{metPotentialThresh}{metPotentialThresh}

\item{currExclude}{currExclude}

\item{uniExclude}{uniExclude}

\item{orphExclude}{orphExclude}
}
\description{
SetMetPotentialOpts
}
