% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{SetSnp2GeneOpts}
\alias{SetSnp2GeneOpts}
\title{SetSnp2GeneOpts}
\usage{
SetSnp2GeneOpts(snpDBType, phescOpt, vepOpt, vepNum, vepDis)
}
\arguments{
\item{snpDBType}{snpDBType}

\item{phescOpt}{phescOpt}

\item{vepOpt}{vepOpt}

\item{vepNum}{vepNum}

\item{vepDis}{vepDis}
}
\description{
SetSnp2GeneOpts
}
