% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkUtils.R
\name{FilterByTissue}
\alias{FilterByTissue}
\title{FilterByTissue}
\usage{
FilterByTissue(dataSetObj = NA, type, tissue)
}
\arguments{
\item{dataSetObj}{dataSetObj}

\item{type}{Input database name}

\item{tissue}{Input the tissue name}
}
\description{
Filter gene/protein by tissue
}
