% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPeakNet.R
\name{InitiatePeakSet}
\alias{InitiatePeakSet}
\title{InitiatePeakSet}
\usage{
InitiatePeakSet(ionMode = -1, ppm = 10, database = "KEGG", p_cutoff = 0.1)
}
\arguments{
\item{ionMode}{ion mode, postive is 1 and negative is -1}

\item{ppm}{ppm value of mass}

\item{database}{database, can be "KEGG", "HMDB" or "Pubchem"}

\item{p_cutoff}{p value cutoff for filtration, default is 0.1}
}
\description{
InitiatePeakSet
}
