% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{PrepareInputList}
\alias{PrepareInputList}
\title{Process input list}
\usage{
PrepareInputList(dataSetObj = "NA", inputList, org, type, queryType)
}
\arguments{
\item{dataSetObj}{Input the name of the created dataSetObj (see Init.Data)}

\item{inputList}{Tab-delimited String of input omics list}

\item{org}{organism code: hsa, mmu, microbiome, rno, cel, dme, dre, sce, bta, gga}

\item{queryType}{}
}
\description{
Process input list
}
