% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFileUtils.R
\name{ReadGraphFile}
\alias{ReadGraphFile}
\title{Title}
\usage{
ReadGraphFile(dataSetObj = NA, fileName, fileType)
}
\arguments{
\item{fileName}{Input name of graph file}

\item{fileType}{File type of graph file (".json", ".graphml", ".txt", ".sif")}
}
\description{
Title
}
